src/requirements.txt
awscli==1.22.30
beautifulsoup4==4.10.0
boto==2.49.0
boto3==1.20.30
botocore==1.23.30
mysql-replication==0.43.0
PyMySQL==1.1.0
python-dateutil==2.8.2
PyYAML==5.4.1
s3transfer==0.5.0
simplejson==3.2.0
virtualenv==20.13.0
